Add temp mds-db dependency to facilitate tests
