+++ packages/mds-db/migration.ts
@@ -158,15 +158,9 @@ async function alterGeographiesColumnsMigration(exec: SqlExecuterFunction) {
 }
 
 async function alterAuditEventsColumnsMigration(exec: SqlExecuterFunction) {
-  await exec(
-    `ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_id bigint`
-  )
-  await exec(
-    `ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_type varchar(31)`
-  )
-  await exec(
-    `ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_type_reason varchar(31)`
-  )
+  await exec(`ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_id bigint`)
+  await exec(`ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_type varchar(31)`)
+  await exec(`ALTER TABLE ${schema.TABLE.audit_events} ADD COLUMN provider_event_type_reason varchar(31)`)
 }
 
 async function alterPreviousGeographiesColumnMigration(exec: SqlExecuterFunction) {
